Linked first commit

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -13,7 +13,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t p;
+	int p;
 
 	p = 0;
